Fixed a potential memory leak in UTF-8 conversions
A potential memory leak was introduced with the back and forth
conversion between "external" UTF8 and Tcl's internal UTF8
variant. The problem shows up when returning multiple tuples, where a
string value larger than TCL_DSTRING_STATIC_SIZE is returned (unless,
this comes from the last column). The problem was induced by an
irregularity passing Tcl_DStrings to functions. The function
Tcl_ExternalToUtfDString() blindly initializes the passed-in pointer
to a Tcl_DString, which will lead to a memory leak, if this
Tcl_DString has already a pre-allocated size.  So, e.g., the usual
idiom setting the length of the dsPtr to 0 to keep the pre-allocated
dynamic memory causes a leak.